Populate package exports and restore GUI
