+++ hspek/__init__.py
@@ -1,0 +1,5 @@
+"""Top-level package for Hspek utilities and GUI."""
+
+from . import logic, gui, utils
+
+__all__ = ["logic", "gui", "utils"]
