+++ hspek/gui/__init__.py
@@ -1,0 +1,5 @@
+"""Graphical user interface for Hspek."""
+
+from .app import TorahTreeApp, HebrewDateSelector
+
+__all__ = ["TorahTreeApp", "HebrewDateSelector"]
