+++ hspek/logic/__init__.py
@@ -1,0 +1,5 @@
+"""Core logic for study scheduling and exports."""
+
+from . import gematria, tree, exporters
+
+__all__ = ["gematria", "tree", "exporters"]
