+++ tests/test_torah_tree.py
@@ -1,83 +1,17 @@
-import importlib.util
-import sys
 import types
-from pathlib import Path
 import pytest
+from pathlib import Path
+
+from hspek.logic import tree, exporters
 
 
 def load_module():
-    path = Path(__file__).resolve().parents[1] / "torah_logic_full_updated.py"
-
-    # Stub external GUI/ICS dependencies so the module can be imported
-    if "customtkinter" not in sys.modules:
-        ctk = types.ModuleType("customtkinter")
-        ctk.set_appearance_mode = lambda *a, **k: None
-        ctk.set_default_color_theme = lambda *a, **k: None
-        ctk.CTk = type("CTk", (), {})
-        ctk.StringVar = lambda *a, **k: None
-        mock_widget = type("MockWidget", (), {})
-        # Provide common widget classes used in the application
-        for name in [
-            "CTkFrame",
-            "CTkButton",
-            "CTkLabel",
-            "CTkRadioButton",
-            "CTkLabelFrame",
-            "CTkCheckBox",
-            "CTkFont",
-        ]:
-            setattr(ctk, name, mock_widget)
-        sys.modules["customtkinter"] = ctk
-
-    if "tkcalendar" not in sys.modules:
-        tkcalendar = types.ModuleType("tkcalendar")
-        tkcalendar.DateEntry = type("DateEntry", (), {})
-        sys.modules["tkcalendar"] = tkcalendar
-
-    if "ics" not in sys.modules:
-        ics = types.ModuleType("ics")
-        ics.Calendar = type("Calendar", (), {})
-        ics.Event = type("Event", (), {})
-        ics.DisplayAlarm = type("DisplayAlarm", (), {})
-        sys.modules["ics"] = ics
-
-    if "tkinter" not in sys.modules:
-        tk = types.ModuleType("tkinter")
-        tk.ttk = types.ModuleType("ttk")
-        tk.filedialog = types.ModuleType("filedialog")
-        tk.messagebox = types.ModuleType("messagebox")
-        sys.modules["tkinter"] = tk
-        sys.modules["tkinter.ttk"] = tk.ttk
-        sys.modules["tkinter.filedialog"] = tk.filedialog
-        sys.modules["tkinter.messagebox"] = tk.messagebox
-
-    if "pyppeteer" not in sys.modules:
-
-        async def fake_launch(*args, **kwargs):
-            class FakePage:
-                async def goto(self, url):
-                    pass
-
-                async def pdf(self, opts):
-                    Path(opts["path"]).write_bytes(b"%PDF-1.4\n%fake")
-
-            class FakeBrowser:
-                async def newPage(self):
-                    return FakePage()
-
-                async def close(self):
-                    pass
-
-            return FakeBrowser()
-
-        fake = types.ModuleType("pyppeteer")
-        fake.launch = fake_launch
-        sys.modules["pyppeteer"] = fake
-
-    spec = importlib.util.spec_from_file_location("torah_tree", path)
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    return module
+    """Provide a unified interface similar to the old monolithic module."""
+    return types.SimpleNamespace(
+        get_length_from_node=tree.get_length_from_node,
+        build_sefaria_ref=tree.build_sefaria_ref,
+        write_bookmark_pdf=exporters.write_bookmark_pdf,
+    )
 
 
 @pytest.fixture(scope="session")
